requirements.txt
anyio==4.1.0
click==8.1.6
fastapi==0.95.1
greenlet==3.0.2
h11==0.14.0
idna==3.6
psycopg2==2.9.9
pydantic==1.10.13
pynest-api==0.1.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.26.1
typing_extensions==4.9.0
uvicorn==0.23.1
